+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Features
+
+- Bump sentry runtime SDK version to `6.1.3` ([#333](https://github.com/getsentry/sentry-android-gradle-plugin/pull/333))
+
 ## 3.1.1
 
 ### Features